requirements.txt
instaloader
rich
schedule
Pillow
tkinter
